# apps/web/hooks/api/model/employeeDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * API
 * Auto-generated API docs
 * OpenAPI spec version: 1.0
 */
import type { BranchDto } from './branchDto';
import type { UserDto } from './userDto';
import type { OwnerDto } from './ownerDto';

export interface EmployeeDto {
  id: string;
  name: string;
  phoneNumber: string;
  email: string;
  branchId: string;
  userId: string;
  ownerId: string;
  branch: BranchDto;
  user: UserDto;
  owner: OwnerDto;
}